Adjusted material for jupyter_notebooks.
