+++ src/ada_topics/jupyter_notebooks/execute_cells/config.py
@@ -1,0 +1,9 @@
+"""Definitions of source files for the current chapter."""
+
+SITE_CONTENTS = {
+    "pages": ("objectives_materials.ipynb",),
+    "other": (
+        # "existing figures etc.",
+    ),
+    "built": ("chapter_template-subchapter_1.pdf",),
+}
