+++ src/ada_topics/jupyter_notebooks/execute_cells/screencast/__init__.py
@@ -1,0 +1,1 @@
+"""Code for producing a screencast."""
