+++ src/ada_topics/jupyter_notebooks/execute_cells/screencast/script.md
@@ -1,0 +1,11 @@
+# Script: Subchapter Name
+
+## First slide
+
+- Create different python cells in which the following operations are performed:
+  - 2+2
+  - a = 2 and return a
+  - return a
+  - a = 3 (highlight that the variable is overwritten)
+  - return a (in a new cell, and show that also the previous one returns 3)
+  - Restart kernel and show that a is not defined
