--- src/ada_topics/jupyter_notebooks/getting_started_with_notebooks_old/config.py
@@ -1,13 +1,0 @@
-"""Definitions of source files for the current chapter."""
-
-SITE_CONTENTS = {
-    "pages": ("objectives_materials.ipynb",),
-    "other": (
-        "screencast/public/activate_and_info.png",
-        "screencast/public/blocked_terminal.png",
-        "screencast/public/folder.png",
-        "screencast/public/notebook.png",
-        "screencast/public/starting_page.png",
-    ),
-    "built": ("python_basics-getting_started_with_notebooks.pdf",),
-}
