--- src/ada_topics/jupyter_notebooks/getting_started_with_notebooks_old/screencast/__init__.py
@@ -1,1 +1,0 @@
-"""Code for producing a screencast."""
