--- src/ada_topics/jupyter_notebooks/getting_started_with_notebooks_old/screencast/script.md
@@ -1,32 +1,0 @@
-# Script: Getting started with notebooks
-
-## Intro page: eCampus
-
-- Work in cloud environment (JupyterHub)
-- Could download everything and work locally
-- Support only for JupyterHub
-
-## Create a new notebook
-
-- Call it `first_steps`
-
-## First cell
-
-`import plotly.express as px`
-
-- Show execute button
-- But execute with `Shift-Enter`
-
-## Second cell
-
-`px.data.gapminder()`
-
-## Execute a few more times
-
-- See how counter increases
-- Powerful & Dangerous
-- Restart kernel, do again
-
-## Always before handing something in
-
-- Restart kernel, run all cells
