--- src/ada_topics/jupyter_notebooks/getting_started_with_notebooks_old/screencast/slides.md
@@ -1,31 +1,0 @@
----
-theme: academic
-class: text-center
-highlighter: shiki
-lineNumbers: false
-info: |
-  ## Applied Data Analytics
-drawings:
-  persist: false
-transition: fade
-title: Applied Data Analytics
-defaults:
-  layout: center
-# themeConfig:
-#   paginationPagesDisabled: true
----
-
-### Applied Data Analytics
-
-<br/>
-
-# Basic Python
-
-### Getting started with notebooks
-
-<br/>
-
-
-Hans-Martin von Gaudecker
-
----
