--- src/ada_topics/jupyter_notebooks/imports/lecture_material.md
@@ -1,35 +1,0 @@
-## Learning objectives
-
-- Students learn about libraries and why they are useful.
-
-- Students learn to import whole libraries and specific functions from libraries.
-
-## Quiz
-
-1. What is a library?
-
-   - A. A collection of books
-   - B. Extensions that come with Python.
-   - C. A collection of python functions that we can use.
-   - D. Python code that someone else has written and made available for us to use.
-
-1. What is the syntax for importing a whole library with an alias?
-
-   - A. `import library as alias`
-   - B. `from library import something`
-   - C. `import library`
-   - D. `from library import *`
-
-1. What is the syntax for importing a specific function from a library?
-
-   - A. `import library as alias`
-   - B. `from library import something`
-   - C. `import library`
-   - D. `from library import *`
-
-1. How do you get the natural log of 3 the `numpy` library?
-
-   - A. `import numpy as np ; np.log(3)`
-   - B. `import numpy as np ; log(3)`
-   - C. `import numpy as np ; numpy.log(3)`
-   - D. `from numpy import log ; log(3)`
