--- src/ada_topics/jupyter_notebooks/imports/script.md
@@ -1,27 +1,0 @@
-# Script: Imports
-
-## Why base python is not enough
-
-- Python is a great language, but it comes with a limited set of built-in functions.
-- You can do sums, divisions, etc. ...
-- But if you want to take the log of a number, there is no built-in function for that.
-
-## How do we solve this?
-
-- We can import libraries that have the tools we need.
-- Libraries are python code that someone else has written and made available for us to
-  use.
-- We can import these libraries into our code and use the functions they contain.
-
-## Example: log function with numpy and math
-
-## Syntax on how to import libraries
-
-- import library as alias
-- from library import something
-
-### Notes to us:
-
-- It seems a bit hard to explain this without explaining functions at all (it's a bit
-  hard to explaining what exactly we import). For now, I would do this implicitly (we
-  use the term but don't tell them exactly what it is).
