+++ src/ada_topics/jupyter_notebooks/markdown_cells/screencast/__init__.py
@@ -1,0 +1,1 @@
+"""Code for producing a screencast."""
