+++ src/ada_topics/jupyter_notebooks/opening_notebooks/config.py
@@ -1,0 +1,9 @@
+"""Definitions of source files for the current chapter."""
+
+SITE_CONTENTS = {
+    "pages": ("objectives_materials.ipynb",),
+    "other": (
+        # "existing figures etc.",
+    ),
+    "built": ("jupyter_notebooks-opening_notebooks.pdf",),
+}
