+++ src/ada_topics/jupyter_notebooks/opening_notebooks/screencast/__init__.py
@@ -1,0 +1,1 @@
+"""Code for producing a screencast."""
