+++ src/ada_topics/jupyter_notebooks/opening_notebooks/screencast/script.md
@@ -1,0 +1,26 @@
+# Script: Opening Notebooks - starting a JupyterHub client from eCampus
+
+## Intro
+
+- Jupyter Notebooks are the main tool for data analytics in Python in the corporate
+  world
+- Notebooks are useful because:
+  - you can have immediate feedback after running a cell
+  - you can include nicely formatted text between python cells (next screencasts).
+- We use JupyterHub. There are other options, but you will be on your own.
+
+## Getting there
+
+- Ecampus page -> JupyterHub -> show contents -> start my server
+- Describe user interface of JupyterLab (files on the left, and launch new things)
+
+## First steps
+
+- To start: launch notebook and rename "first_steps". Describe extension ipynb
+- Fill a cell with print("hello world") (mentioning that they don't need to understand
+  the content) and show how to execute it
+  - run cell
+  - Shift-Enter
+  - ctrl-Enter
+- Create a new cell and show that you can move the order of cells - but explain that
+  this is not common.
