+++ src/ada_topics/python_basics/assignment_and_scalar_types/screencast/lecture_material.md
@@ -1,0 +1,69 @@
+# Learning Objectives
+
+Two screencasts out of this:
+
+_(assigning)_
+
+- Assign variables in Python
+- Do simple manipulations with variables
+- Compare variables
+
+would be without much background, e.g. a > 5 or c = b \< a, see output True/False
+without thinking much about it
+
+_(types)_
+
+- List several built-in types in Python _(float, int, Boolean, string)_
+- Explain why not all operations are possible with all types
+
+# Script
+
+[Old one](https://electure.uni-bonn.de/static/mh_default_org/engage-player/6aa0b00c-298f-48d9-a9ad-afc5ccb865ac/872f2264-bf69-4d43-b767-512b0f8ca291/8682e6e5-462b-4ab2-bf3d-f2085d723978.mp4)
+is good.
+
+# Additional Materials
+
+[Documentation](https://docs.python.org/3/library/stdtypes.html) of Python built-in
+types
+
+# Quiz
+
+If you need to represent a real number, in Python you would use ...
+
+- F Boolean
+- F Int
+- T Float
+
+Which of the following are properties about floats?
+
+- F The can represent any real number in a given interval
+- F They can represent arbitrarily large numbers
+- T They can be used for scientific calculations
+- T They can be added to integers
+
+Of which type is 1.0?
+
+- F Boolean
+- F Int
+- T Float
+
+How can you calculate the square of the variable "a"?
+
+- F a^2
+- F a\*\*a
+- T a\*a
+- T a\*\*2
+
+Which values can a Boolean take?
+
+- True
+- False
+- false
+- true
+
+If a = True and b = False, then
+
+- T not a is False
+- F a and b is True
+- F a and not b is False
+- T a or b is True
